requirements.txt
Django==4.1.4
psycopg2-binary==2.9.5
djangorestframework==3.14.0
django-filter==22.1
asttokens==2.0.8
attrs==21.4.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
commonmark==0.9.1
decorator==5.1.1
defusedxml==0.7.1
executing==0.10.0
idna==3.3
iniconfig==1.1.1
IPy==1.1
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
multidict==6.0.2
mypy==0.982
mypy-extensions==0.4.3
numpy==1.23.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
Pygments==2.13.0
pyparsing==3.0.8
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.4
requests==2.27.1
requests-toolbelt==0.10.0
rich==12.6.0
six==1.16.0
stack-data==0.4.0
tomli==2.0.1
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.9
wcwidth==0.2.5
